[수정]헤더 ,중복코드
