+++ Hospital ERP/src/main/java/hospital/service/MemberForgotPwProAction.java
@@ -10,7 +10,7 @@ import javax.servlet.http.HttpSession;
 
 import hospital.action.Action;
 import hospital.action.ActionForward;
-import hospital.dao.Member2Dao;
+import hospital.dao.MemberDao;
 import hospital.utils.NaverMailSend;
 import hospital.vo.LoginVO;
 
@@ -31,20 +31,22 @@ public class MemberForgotPwProAction implements Action {
 			PrintWriter out;
 			try {
 				out = resp.getWriter();
-				out.println("<script>" + "alert('접근 권한이 없습니다.1');" + "location.href='/login.do';" + "</script>");
+				out.println("<script>" + "alert('접근 권한이 없습니다.');" + "location.href='/login.do';" + "</script>");
 				out.flush();
 				out.close();
 			} catch (IOException e) {
 				e.printStackTrace();
 			}
 			return null;
+
 		} else {
-			
+
 			String h_email = req.getParameter("h_email");
 			String h_id = req.getParameter("h_id");
 
-			Member2Dao dao = new Member2Dao();
+			MemberDao dao = new MemberDao();
 			LoginVO vo = (LoginVO) dao.getMember(h_id, h_email);
+			
 			if (vo == null || !vo.getH_id().equals(h_id) || !vo.getH_email().equals(h_email)) {
 				try {
 					resp.sendError(404);
@@ -52,14 +54,13 @@ public class MemberForgotPwProAction implements Action {
 					e.printStackTrace();
 				}
 			} else {
-				
+
 				NaverMailSend mailSend = new NaverMailSend();
 
 				try {
 					String authenticationCode = mailSend.sendEmail(h_email);
 					session.setAttribute("authCode", authenticationCode);
 					session.setAttribute("id", h_id);
-					
 
 					forward = new ActionForward();
 					forward.setPath("./memberChangePw.do");
