+++ Hospital ERP/src/main/java/hospital/vo/LoginVO.java
@@ -1,7 +1,5 @@
 package hospital.vo;
 
-import java.sql.Connection;
-
 //로그인 정보
 public class LoginVO {
